r671: tweaks to make the eDiss suite a bit more functional:
- you can now use a VFC like "2.1.x" or "6.x.x"
- all etyma, sequenced or not, are included if they match the sought-after VFC
- online table of contents in ~stedt/public_html/dissemination is now properly updated.

The dang thing hardly worked before! Very marginal!

+++ printutils/makeFasc.sh
@@ -9,17 +9,23 @@
 # if --i is specified, the draft format is created and moved to the site, 
 # but the ToC is not updated.
 #
+if [ -z $1 ] ; then echo need a value for V; exit ; fi
+if [ -z $2 ] ; then echo need a value for F; exit ; fi
+if [ -z $3 ] ; then echo need a value for C; exit ; fi
 # first, get to the right place
 if [ "$4" = '--i' ] ; then
 DRAFT="-draft"
 else
 DRAFT=""
 fi
+# verbose...
 set -x
 cd ~stedt-cgi-ssl/rootcanals/
 cd tex/
 rm $1-$2-$3.*
 cd ..
+# exit on errors..
+set -e
 # generate the .tex file
 perl extract.pl $1 $2 $3 $4
 cd tex/
@@ -38,7 +44,8 @@ cp $pdffile.pdf ~stedt/public_html/dissemination/$pdffile-$DATETIME-1$DRAFT.pdf
 if [ "$4" = '--i' ] ; then
 echo "done with *DRAFT* $texfile"
 else 
-perl ~stedt-cgi-ssl/rootcanals/makeToC.pl > ~stedt/public_html/dissemination.html       
 echo "done with $texfile" 
 fi
+perl ~stedt-cgi-ssl/rootcanals/makeToC.pl > ~stedt/public_html/dissemination.html       
+echo "eDiss table of contents updated."
 echo "http://stedt.berkeley.edu/dissemination/$pdffile-$DATETIME-1$DRAFT.pdf"
